+++ bundles/tampere/content-editor/index.js
@@ -1,0 +1,7 @@
+import './instance'
+import './resources/scss/style.scss'
+
+
+Oskari.bundle('content-editor', () =>
+    Oskari.clazz.create('Oskari.tampere.bundle.content-editor.ContentEditorBundleInstance')
+);