+++ bundles/tampere/selected-featuredata/index.js
@@ -1,0 +1,7 @@
+import './instance'
+import './Flyout'
+import './resources/css/style.css'
+
+Oskari.bundle('selected-featuredata', () =>
+    Oskari.clazz.create('Oskari.mapframework.bundle.selected-featuredata.SelectedFeatureDataBundleInstance')
+);