+++ api/CHANGELOG.md
@@ -1,3 +1,12 @@
 # Contrib API changelog
 
 This document describes changes to the public API of contrib bundles
+
+# > 3.2.0
+
+Migrated bundles:
+
+ * packages/tampere/bundle/admin-wfs-search-channel/bundle.js -> bundles/tampere/admin-wfs-search-channel
+ * packages/tampere/bundle/content-editor/bundle.js ->  bundles/tampere/content-editor 
+ * packages/tampere/bundle/search-from-channels/bundle.js ->  bundles/tampere/search-from-channels
+ * packages/tampere/bundle/selected-featuredata/bundle.js ->  bundles/tampere/selected-featuredata